adds click events to the songs that you can search for.
once you load the songs you can search for into the page this adds a click
event to each song and pulls the data-id from the song and puts it in the
console log. Issue i am having is exporting the click handlers callback
function to the song-events file where the ajax call would be more
appropriate.