this adds the edit feature onto the your queue page
this sends a patch request to the api to update the song you decide to
edit. right now I have it hard coded to update to song number one no
matter what. need to build the functionality to choose which song to
update it to. Also does not change on the ui yet just changes it in the
back end