builds out the feature to allow the user to edit what song they want to
before it was hard coded to edit to song id 1. this builds in a new get
request into the your list view to show all the songs to choose from.
then based on what song is choosen from the get list i used its id to send
into the patch request and update the ui with handlebars