fixes the hitting the edit button twice issue. moves headers into the view
Every time you clicked on edit it made a call and kept inserting another
song list view underneath the one someone already had open. I fixed that
by adding a if edit view is visible statement to handle that. also added
the headers into the handlebars files instead of them being hard coded
in the html so that the headers lined up better and the view is all one
element